Add shell script linting: Check for shellcheck warnings in shell scripts
Summary:
Adapted to run as an Arcanist linter, and added fixes for the scripts
which are Bitcoin ABC only.

Because the quantity of shell script is limited, the performance cost is
almost unnoticeable.

Backport of core PR12871
https://github.com/bitcoin/bitcoin/pull/12871/files

Depends on D3956, D3957 and D3958.

Test Plan:
Should return no error:
  arc lint --everything

Add an unsused variable in a script and lint that file:
  arc lint -- <the_script_file>
Check the linter returns a warning.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D3959

+++ contrib/teamcity/build.sh
@@ -39,7 +39,7 @@ SAN_LOG_DIR="${BUILD_DIR}/sanitizer_logs"
 
 # Create the log directory if it doesn't exist and clear it
 mkdir -p "${SAN_LOG_DIR}"
-rm -rf "${SAN_LOG_DIR}"/*
+rm -rf "${SAN_LOG_DIR:?}"/*
 
 # Sanitizers options, not used if sanitizers are not enabled
 export ASAN_OPTIONS="malloc_context_size=0:log_path=${SAN_LOG_DIR}/asan.log"
